pyproject.toml
[tool.poetry]
name = "django-basic"
version = "0.1.0"
description = ""
authors = ["Steve Jones <jonesst2608@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11.4"
django = "^5.0.6"
django-unfold = "^0.25.0"
reactpy-django = "^3.8.1"
channels = {extras = ["daphne"], version = "^4.1.0"}
convert2reactpy = {version = "^0.0.6", source = "pypicloud"}

[tool.poetry.dev-dependencies]
pylint-django = "^2.5.5"

[[tool.poetry.source]]
name = "pypicloud"
url = "http://debian-server:6543/simple/"
priority = "supplemental"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
